Add RestManager class and refactor networking code
