+++ WorldWeather/Controller/Second tab/SearchLocationViewController.swift
@@ -18,7 +18,8 @@ class SearchLocationViewController: UIViewController {
         
         //searchLocationView:.segmentedControl.selectedSegmentIndex = defaults.integer(forKey: "temperatureUnit")
     }
-
+    
+    
 
 }
